Remove specialized generator-iterator classes
Also split out of DataDistribution.h (into BlockGenerator.h) the
generator-related classes
