+++ src/DataDistribution.h
@@ -5,7 +5,6 @@
 #include <algorithm>
 #include <cassert>
 #include <cstddef>
-#include <functional>
 #include <iterator>
 #include <memory>
 #include <optional>
@@ -13,6 +12,8 @@
 #include <type_traits>
 #include <tuple>
 #include <vector>
+
+#include <BlockGenerator.h>
 
 namespace mpims {
 
@@ -55,276 +56,7 @@ blocks(const InputIterator& first, const InputIterator& last) {
   return result;
 }
 
-// BlockGeneratorIterator
-//
-// index iterator from index block generator function
-//
-template <typename S>
-class BlockGeneratorIterator
-  : public std::iterator<std::input_iterator_tag, std::size_t> {
-
-public:
-
-  typedef std::function<
-  std::tuple<
-    S,
-    std::optional<std::tuple<std::size_t, std::size_t> > >(const S&)>
-  generator_t;
-
-  BlockGeneratorIterator(
-    const generator_t& generator,
-    S&& state,
-    bool at_end = false)
-    : m_generator(generator)
-    , m_block_offset(0) {
-    if (!at_end)
-      std::tie(m_next_state, m_block) = m_generator(state);
-  }
-
-  BlockGeneratorIterator(const BlockGeneratorIterator& other)
-    : m_generator(other.m_generator)
-    , m_next_state(other.m_next_state)
-    , m_block(other.m_block)
-    , m_block_offset(other.m_block_offset) {
-  }
-
-  bool
-  operator==(const BlockGeneratorIterator& rhs) {
-    if (at_end() && rhs.at_end())
-      return true;
-    if (at_end() || rhs.at_end())
-      return false;
-    return (m_block == rhs.m_block
-            && m_block_offset == rhs.m_block_offset
-            && m_generator == rhs.m_generator);
-  }
-
-  bool
-  operator!=(const BlockGeneratorIterator& rhs) {
-    return !operator==(rhs);
-  }
-
-  std::size_t
-  operator*() const {
-    std::size_t b0;
-    std::tie(b0, std::ignore) = m_block.value();
-    return b0 + m_block_offset;
-  }
-
-  BlockGeneratorIterator&
-  operator++() {
-    std::size_t b0, blen;
-    std::tie(b0, blen) = m_block.value();
-    if (++m_block_offset == blen) {
-      std::tie(m_next_state, m_block) = m_generator(m_next_state);
-      assert(!m_block || std::get<0>(m_block.value()) >= b0 + blen);
-      m_block_offset = 0;
-    }
-    return *this;
-  }
-
-  BlockGeneratorIterator
-  operator++(int) {
-    BlockGeneratorIterator result(*this);
-    operator++();
-    return result;
-  }
-
-  bool
-  at_end() const {
-    return !m_block;
-  }
-
-  std::vector<std::size_t>
-  take(std::size_t n = 1) {
-    std::vector<std::size_t> result;
-    result.resize(n);
-    while (n > 0 && !at_end()) {
-      result.push_back(operator*());
-      operator++();
-    }
-    return result;
-  }
-
-protected:
-
-  void
-  swap(BlockGeneratorIterator& other) {
-    using std::swap;
-    swap(m_generator, other.m_generator);
-    swap(m_next_state, other.m_next_state);
-    swap(m_block, other.m_block);
-    swap(m_block_offset, other.m_block_offset);
-  }
-
-private:
-
-  generator_t m_generator;
-  S m_next_state;
-  std::optional<std::tuple<std::size_t, std::size_t> > m_block;
-  std::size_t m_block_offset;
-};
-
-// CyclicGenerator
-//
-// generator function for cyclic distribution of index blocks
-//
-class CyclicGenerator {
-
-public:
-
-  struct State {
-    std::size_t axis_length;
-    std::size_t block_length;
-    std::size_t group_length;
-    std::size_t index;
-  };
-
-  static std::tuple<
-    State,
-    std::optional<std::tuple<std::size_t, std::size_t> > >
-  apply(const State& st) {
-
-    auto index = st.index + st.group_length;
-    if (index < st.axis_length) {
-      State next_st {st.axis_length, st.block_length, st.group_length, index};
-      auto blen = std::min(st.axis_length - index, st.block_length);
-      return std::make_tuple(next_st, std::make_tuple(index, blen));
-    } else {
-      State next_st {
-        st.axis_length, st.block_length, st.group_length, st.axis_length};
-      return std::make_tuple(next_st, std::nullopt);
-    }
-  }
-};
-
-// CyclicIterator
-//
-// index iterator for block-cyclic distribution
-//
-class CyclicIterator
-  : public BlockGeneratorIterator<CyclicGenerator::State> {
-
-public:
-
-  CyclicIterator(
-    std::size_t axis_length,
-    std::size_t offset,
-    std::size_t block_length,
-    std::size_t group_size,
-    bool at_end = false)
-    : BlockGeneratorIterator(
-      CyclicGenerator::apply,
-      CyclicGenerator::State{
-        axis_length,
-          block_length,
-          block_length * group_size,
-          at_end ? axis_length : offset},
-      at_end) {
-  }
-};
-
-// BlockSequenceGenerator
-//
-// generator function for sequence (vector) of blocks
-//
-class BlockSequenceGenerator {
-
-public:
-
-  struct State {
-    std::vector<std::tuple<std::size_t, std::size_t> > blocks;
-    std::size_t block_index;
-  };
-
-  static std::tuple<
-    State,
-    std::optional<std::tuple<std::size_t, std::size_t> > >
-  apply(const State& st) {
-
-    auto blk = st.block_index + 1;
-    State next_st {st.blocks, blk};
-    if (blk < st.blocks.size())
-      return std::make_tuple(next_st, st.blocks[blk]);
-    else
-      return std::make_tuple(next_st, std::nullopt);
-  }
-};
-
-// BlockSequenceIterator
-//
-// index iterator for sequence of blocks
-//
-class BlockSequenceIterator
-  : public BlockGeneratorIterator<BlockSequenceGenerator::State> {
-
-public:
-
-  BlockSequenceIterator(
-    const std::vector<std::tuple<std::size_t, std::size_t> >& blocks,
-    bool at_end = false)
-    : BlockGeneratorIterator(
-      BlockSequenceGenerator::apply,
-      BlockSequenceGenerator::State {blocks, at_end ? blocks.size() : 0},
-      at_end) {
-  }
-};
-
-// BlockGenerator
-//
-// generator function for block-valued iterator
-//
-template <
-  typename InputIterator,
-  class = typename std::enable_if<
-    std::is_convertible<
-      typename std::iterator_traits<InputIterator>::value_type,
-      std::tuple<std::size_t, std::size_t> >::value>::type>
-class BlockGenerator {
-
-public:
-  struct State {
-    InputIterator current;
-    InputIterator end;
-  };
-
-  static std::tuple<
-    State,
-    std::optional<std::tuple<std::size_t, std::size_t> > >
-  apply(const State& st) {
-
-    auto next(st.current);
-    ++next;
-    State next_st {next, st.end};
-    if (next != st.end)
-      return std::make_tuple(next_st, *next);
-    else
-      return std::make_tuple(next_st, std::nullopt);
-  }
-};
-
-// BlockIterator
-//
-// index iterator for block-yielding iterator
-//
-template <
-  typename InputIterator,
-  class = typename std::enable_if<
-    std::is_convertible<
-      typename std::iterator_traits<InputIterator>::value_type,
-      std::tuple<std::size_t, std::size_t> >::value>::type>
-class BlockIterator
-  : public BlockGeneratorIterator<typename BlockGenerator<InputIterator>::State> {
-
-public:
-
-  BlockIterator(InputIterator&& begin, InputIterator&& end)
-    : BlockGeneratorIterator<BlockGenerator<InputIterator> >(
-      BlockGenerator<InputIterator>::apply,
-      typename BlockGenerator<InputIterator>::State {begin, end},
-      begin == end) {
-  }
-};
+class DataDistributionFactory;
 
 // DataDistribution
 //
@@ -411,7 +143,7 @@ public:
 // base class for all sub-classes of DataDistribution defined in this module
 //
 template <
-  typename T, 
+  typename T,
   class = typename std::enable_if<
     std::is_convertible<
       typename std::iterator_traits<T>::value_type,
@@ -422,18 +154,6 @@ class IteratorDataDistribution
 public:
 
   typedef T iterator;
-
-  IteratorDataDistribution(iterator&& begin)
-    : m_begin(std::forward<iterator>(begin)) {
-  }
-
-  IteratorDataDistribution(const IteratorDataDistribution& other)
-    : m_begin(other.m_begin) {
-  }
-
-  IteratorDataDistribution(IteratorDataDistribution&& other)
-    : m_begin(std::move(other).m_begin) {
-  }
 
   IteratorDataDistribution&
   operator=(const IteratorDataDistribution& other) {
@@ -453,11 +173,12 @@ public:
   class IteratorDataDistributionIterator
     : public Iterator {
 
-  public:
+    friend class IteratorDataDistribution<T>;
 
-    IteratorDataDistributionIterator(const T& iter)
-      : m_iter(iter) {
-    }
+    friend std::unique_ptr<IteratorDataDistributionIterator>
+    std::make_unique<IteratorDataDistributionIterator>(const T&);
+
+  public:
 
     std::size_t
     operator*() const override {
@@ -473,6 +194,12 @@ public:
     bool
     at_end() const override {
       return m_iter.at_end();
+    }
+
+  protected:
+
+    IteratorDataDistributionIterator(const T& iter)
+      : m_iter(iter) {
     }
 
   private:
@@ -508,6 +235,18 @@ public:
 
 protected:
 
+  IteratorDataDistribution(iterator&& begin)
+    : m_begin(std::forward<iterator>(begin)) {
+  }
+
+  IteratorDataDistribution(const IteratorDataDistribution& other)
+    : m_begin(other.m_begin) {
+  }
+
+  IteratorDataDistribution(IteratorDataDistribution&& other)
+    : m_begin(std::move(other).m_begin) {
+  }
+
   iterator
   itbegin() const {
     return m_begin;
@@ -516,7 +255,7 @@ protected:
   void
   swap(IteratorDataDistribution& other) {
     m_begin.swap(other.m_begin);
-    
+
   }
 
 private:
@@ -526,92 +265,47 @@ private:
 
 // BlockGeneratorDataDistribution
 //
-// data distribution for a BlockGeneratorIterator type
+// data distribution for a generator
 //
-// various sub-types of BlockGeneratorIterator defined in this module also have
-// direct sub-classes of IteratorDataDistribution defined here, but this
-// provides a generic definition without any sort of specialized name or
-// constructors
+// this is the implementation sub-class of DataDistribution, which should
+// support almost all sorts of data distributions (all that's required is a
+// generator function and an initial state)
 //
 template <typename S>
 class BlockGeneratorDataDistribution
   : public IteratorDataDistribution<BlockGeneratorIterator<S> > {
 
-public:
+  typedef typename BlockGeneratorIterator<S>::generator_t generator_t;
 
-  BlockGeneratorDataDistribution(
-    const typename BlockGeneratorIterator<S>::generator_t& generator,
-    const S& state)
+  friend std::unique_ptr<BlockGeneratorDataDistribution>
+  std::make_unique<BlockGeneratorDataDistribution>(
+    const generator_t& generator,
+    S&&);
+
+private:
+
+  BlockGeneratorDataDistribution(const generator_t& generator, S&& state)
     : IteratorDataDistribution<BlockGeneratorIterator<S> >(
-      BlockGeneratorIterator(generator, state)) {
+      BlockGeneratorIterator(generator, std::forward<S>(state))) {
   }
-};
-
-// CyclicDataDistribution
-//
-// block-cyclic data distribution
-//
-class CyclicDataDistribution
-  : public IteratorDataDistribution<CyclicIterator> {
 
 public:
 
-  CyclicDataDistribution(
-    std::size_t axis_length,
-    std::size_t offset,
-    std::size_t block_length,
-    std::size_t group_size)
-    : IteratorDataDistribution(
-      CyclicIterator(
-        axis_length,
-        offset,
-        block_length,
-        group_size)) {
+  static std::unique_ptr<DataDistribution>
+  make(const generator_t& generator, S&& state) {
+    return std::make_unique<BlockGeneratorDataDistribution>(
+      generator,
+      std::forward<S>(state));
   }
-};
 
-// BlockSequenceDataDistribution
-//
-// (enumerated) block sequence data distribution
-//
-class BlockSequenceDataDistribution
-  : public IteratorDataDistribution<BlockSequenceIterator> {
-
-public:
-
-  BlockSequenceDataDistribution(
-    const std::vector<std::tuple<std::size_t, std::size_t> >& blocks)
-    : IteratorDataDistribution(BlockSequenceIterator(blocks)) {
-  }
-};
-
-// BlockIteratorDataDistribution
-//
-// block iterator data distribution
-//
-template <
-  typename InputIterator,
-  class = typename std::enable_if<
-    std::is_convertible<
-      typename std::iterator_traits<InputIterator>::value_type,
-      std::tuple<std::size_t, std::size_t> >::value>::type>
-class BlockIteratorDataDistribution
-  : public IteratorDataDistribution<BlockIterator<InputIterator> > {
-
-public:
-
-  BlockIteratorDataDistribution(InputIterator&& begin, InputIterator&& end)
-    : IteratorDataDistribution<BlockIterator<InputIterator> >(
-      BlockIterator(
-        std::forward<InputIterator>(begin),
-        std::forward<InputIterator>(end))) {
-  }
 };
 
 class DataDistributionFactory {
 
 public:
 
+  // block-cyclic data distribution
+  //
   static std::unique_ptr<DataDistribution>
   cyclic(
     std::size_t axis_length,
@@ -619,21 +313,30 @@ public:
     std::size_t block_length,
     std::size_t group_size) {
 
-  return
-    std::make_unique<CyclicDataDistribution>(
-      axis_length,
-      offset,
-      block_length,
-      group_size);
+    return
+      BlockGeneratorDataDistribution<CyclicGenerator::State>::make(
+        CyclicGenerator::apply,
+        CyclicGenerator::State{
+          axis_length,
+            block_length,
+            block_length * group_size,
+            std::min(offset, axis_length)});
   }
 
+  // (enumerated) block sequence data distribution
+  //
   static std::unique_ptr<DataDistribution>
   block_sequence(
     const std::vector<std::tuple<std::size_t, std::size_t> >& blocks) {
 
-    return std::make_unique<BlockSequenceDataDistribution>(blocks);
+    return
+      BlockGeneratorDataDistribution<BlockSequenceGenerator::State>::make(
+        BlockSequenceGenerator::apply,
+        BlockSequenceGenerator::State{blocks, 0});
   }
-  
+
+  // block iterator data distribution
+  //
   template <
     typename InputIterator,
     class = typename std::enable_if<
@@ -644,21 +347,24 @@ public:
   block_iterator(InputIterator&& begin, InputIterator&& end) {
 
     return
-      std::make_unique<BlockIteratorDataDistribution>(
-        std::forward<InputIterator>(begin),
-        std::forward<InputIterator>(end));
+      BlockGeneratorDataDistribution<
+        typename BlockGenerator<InputIterator>::State>::make(
+          BlockGenerator<InputIterator>::apply,
+          typename BlockGenerator<InputIterator>::State{begin, end});
   }
 
+  // generic data distribution factory method
+  //
   template <typename S>
   static std::unique_ptr<DataDistribution>
   block_generator(
     const typename BlockGeneratorIterator<S>::generator_t& generator,
-    const S& init_state) {
+    S&& init_state) {
 
     return
-      std::make_unique<BlockGeneratorIterator<S> >(
+      BlockGeneratorDataDistribution<S>::make(
         generator,
-        init_state);
+        std::forward<S>(init_state));
   }
 };
